changed int type to int64_t

+++ 6_2_1/ex1/include/custom_int.hpp
@@ -1,26 +1,27 @@
 #ifndef CUSTOM_INT_H
 #define CUSTOM_INT_H
+#include <cstdint>
 class CustomInt {
 public:
   CustomInt();
   ~CustomInt();
-  int& operator=(CustomInt const &obj);
-  int& operator=(const int val);
-  int& operator+(CustomInt const &obj);
-  int& operator+(const int val);
-  int& operator+=(CustomInt const &obj);
-  int& operator+=(const int val);
-  int& operator-(CustomInt const &obj);
-  int& operator-(const int val);
-  int& operator-=(CustomInt const &obj);
-  int& operator-=(const int val);
+  int64_t& operator=(CustomInt const &obj);
+  int64_t& operator=(const int val);
+  int64_t& operator+(CustomInt const &obj);
+  int64_t& operator+(const int val);
+  int64_t& operator+=(CustomInt const &obj);
+  int64_t& operator+=(const int val);
+  int64_t& operator-(CustomInt const &obj);
+  int64_t& operator-(const int val);
+  int64_t& operator-=(CustomInt const &obj);
+  int64_t& operator-=(const int val);
 
   void set_value(const int value);
-  const int& get_value()const;
+  const int64_t& get_value()const;
   void print()const;
   void print_counter_value() const;
 private:
-  int m_value;
+  int64_t m_value;
   static int counter;
 };
 #endif
